Added View button to simply retrieve info from DB Logging in clears text in login screen fields Certain fields are either editable or not, depending on whether an HR person is logged in or a regular employee
